refactor: Slider.jsx
