+++ src/components/Slider/slider.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import propTypes from 'prop-types';
 import { getNextRoundRobin, getRandomNumber } from '../../libs/utils/math';
-import { PUBLIC_IMAGE_FOLDER, DEFAULT_BANNER_IMAGE } from './configs/constants';
+import { PUBLIC_IMAGE_FOLDER, DEFAULT_BANNER_IMAGE } from '../../configs/constants';
 
 import Img from './style';
 
@@ -9,26 +9,22 @@ export default class Slider extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      path: '',
       index: 0,
     };
   }
 
   componentDidMount() {
     const {
-      banners, duration, random, defaultbanner,
+      banners, duration, random,
     } = this.props;
-    this.id = setInterval(() => {
+    this.setSliderInterval = setInterval(() => {
       const { index } = this.state;
       let imageIndex = 0;
-      let imagePath = `${defaultbanner}`;
-      if (banners.length) {
+      if (banners && banners.length) {
         imageIndex = random ? getRandomNumber(banners.length)
           : getNextRoundRobin(banners.length, index);
-        imagePath = `${PUBLIC_IMAGE_FOLDER}${banners[imageIndex]}`;
       }
       this.setState({
-        path: imagePath,
         index: imageIndex,
       });
     }, duration);
@@ -36,21 +32,15 @@ export default class Slider extends React.Component {
 
 
   componentWillUnmount() {
-    clearInterval(this.id);
+    clearInterval(this.setSliderInterval);
   }
 
   render() {
-    const { path, index } = this.state;
+    const { index } = this.state;
     const {
-      height, altText, banners,
+      height, altText, banners, defaultbanner,
     } = this.props;
-
-    let imgPath;
-    if (!path) {
-      imgPath = banners.length ? `${PUBLIC_IMAGE_FOLDER}${banners[index]}` : DEFAULT_BANNER_IMAGE;
-    } else {
-      imgPath = path;
-    }
+    const imgPath = banners.length ? `${PUBLIC_IMAGE_FOLDER}${banners[index]}` : `${defaultbanner}`;
     return (
       <>
         <Img
